requirements.txt
git+https://github.com/ivanmkc/rasa.git@dev
rasa-sdk>=2.5.0  # if you change this, make sure to change the Dockerfile to match
spacy
pyyaml
-r actions/requirements-actions.txt